@@ -99,11 +99,11 @@
     "prettier": "^1.11.1",
     "promise": "8.0.1",
     "raf": "3.4.0",
-    "react": "^16.2.0",
+    "react": "^16.3.2",
     "react-codemirror2": "^3.0.7",
     "react-dev-utils": "^5.0.0",
     "react-document-title": "^2.0.3",
-    "react-dom": "^16.2.0",
+    "react-dom": "^16.3.2",
     "react-live": "^1.10.1",
     "react-redux": "^5.0.7",
     "react-router": "^4.2.0",
